backend/requirements.txt
# ALERTA-LINK Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (opcional, para produccion)
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# ML/Data
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0

# HTTP client (para crawler opcional)
httpx>=0.26.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# WHOIS lookup (para verificar antiguedad de dominios)
python-whois>=0.9.0

# Crawler headless (para analisis profundo de URLs)
playwright>=1.40.0

# Security
cryptography>=41.0.0

# Rate Limiting
slowapi>=0.1.9

# HTTP Requests
requests>=2.31.0
